refactor(client): split websocket store into domain event modules
Split 2368-line stores/websocket.ts into stores/websocket/ with index.ts
(connection state, lifecycle, dispatch table) plus six domain handler
modules: messageEvents (edit + thread + notification), voiceEvents
(SDP/ICE/participants/stats/layers), mediaEvents (screen share + webcam),
adminEvents (bans/suspensions/deletions/reports), guildEvents (emoji +
patch), reactionEvents (reactions + channel pin message status).

Existing `@/stores/websocket` imports continue to work; media handlers
are re-exported from index.ts for test compatibility.
